// packages/ooxml-viewer/src/openxml/Types.ts
export type ST_Em = 'none' | 'dot' | 'comma' | 'circle' | 'underDot';
// ...
export type ST_TextAlignment =
  | 'top'
  | 'center'
  | 'baseline'
  | 'bottom'
  | 'auto';